feat(registry): extract brand asset helpers into brand-assets-lib
Move pure domain normalization, known-brand detection, Brandfetch/proxy
URL construction, and metadata sanitization out of brand-assets.js into
brand-assets-lib.js. The public brand-assets.js module and
@heyclaude/registry/brand-assets export now re-export that surface so
existing imports stay unchanged.

Add brand-assets-lib unit tests covering hosting-domain guards, alias
matching, client-id resolution, and metadata sanitization — 100% branch
coverage for the new lib.
